chore(cypress-ete): split `registration_1.8.cy.ts` into individual tests
